requirements_enhanced.txt
# Enhanced VLM3D Task 2 Requirements
# Core ML dependencies (should already be installed)
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.8.0
torchmetrics>=0.10.0

# Data processing
numpy>=1.21.0
pandas>=1.4.0
opencv-python>=4.6.0
pillow>=9.0.0
nibabel>=4.0.0
scikit-image>=0.19.0
scipy>=1.8.0

# Compression libraries
lz4>=4.0.0

# Visualization and logging
matplotlib>=3.5.0
tensorboard>=2.9.0
tqdm>=4.64.0

# ML utilities
scikit-learn>=1.1.0
huggingface-hub>=0.10.0

# For enhanced features
timm>=0.6.0  # For additional model architectures if needed
albumentations>=1.3.0  # Alternative augmentation library (optional)
